Add files via upload
The scenario is to calculate how much a logistics company has overpaid, properly paid and underpaid their freight contractor, provided different weight slabs for packages, and also consider the delivery distance. This project resembles the recent development of e-commerce and its delivery system and how the proper calculation is important while we work with the external contracts.